html_format {markdown}	R Documentation

R Markdown output formats

Description

Convenience functions for R Markdown v2 users.

Usage


html_format(
  meta = NULL,
  template = NULL,
  options = NULL,
  keep_md = FALSE,
  keep_tex = FALSE,
  latex_engine = "xelatex"
)

latex_format(
  meta = NULL,
  template = NULL,
  options = NULL,
  keep_md = FALSE,
  keep_tex = FALSE,
  latex_engine = "xelatex"
)

Arguments

meta, template, options
Arguments to be passed to mark().

keep_md, keep_tex
Whether to keep the intermediate ‘.md’ and ‘.tex’ files generated from ‘.Rmd’.

latex_engine
The LaTeX engine to compile ‘.tex’ to ‘.pdf’. This argument and keep_tex are for latex_format() only, and ignored in html_format().

Details

We refer to this markdown package plus knitr as “R Markdown v1”, and the rmarkdown package as “R Markdown v2”. The former uses commonmark to convert Markdown, and the latter uses Pandoc. However, the latter also accept custom converting tools in addition to Pandoc. The output formats here provide the custom converting function mark() to rmarkdown, so that users can take advantage of rmarkdown::render() and the Knit button in RStudio. It is absolutely not necessary to rely on rmarkdown. The only point is convenience. If you do not use rmarkdown::render() or the Knit button, you can definitely just call markdown::mark() directly.

[Package markdown version 1.12 Index]