markdown {markdown}	R Documentation

Markdown rendering for R

Description

Markdown is a plain-text formatting syntax that can be converted to XHTML or other formats. This package provides R bindings to the Sundown (https://github.com/vmg/sundown) markdown rendering library.

Details

The R function markdownToHTML renders a markdown file to HTML (respecting the specified markdownExtensions and markdownHTMLOptions).

The package also exports the underlying Sundown C extension API which enables creating and calling custom renderers using the renderMarkdown function.

To learn more about markdown syntax see:

http://en.wikipedia.org/wiki/Markdown

Author(s)

JJ Allaire, Jeffrey Horner, Vicent Marti, and Natacha Porte

Maintainer: Yihui Xie <xie@yihui.name>

See Also

markdownToHTML renderMarkdown

[Package markdown version 1.0 Index]