registeredRenderers {markdown}	R Documentation

List of Registered Markdown Renderers

Description

registeredRenderers returns a named character vector listing all the registered renderers known to the markdown package. markdown allows up to seven renderers to be registered by users; HTML is provided by the package.

Usage

registeredRenderers()

Value

A named character vector listing all available renderers. Vector value contain renderer names, and named values contain the renderer output type, either character or raw.

See Also

markdownToHTML, rendererOutputType

Examples

# List all available renderers
registeredRenderers()

[Package markdown version 1.0 Index]