renderMarkdown {markdown}	R Documentation

Functions in markdown before v1.3.

Description

These functions are kept in this package for backward-compatibility. They will not be removed in the foreseeable future, although we recommend that you use their new names instead: renderMarkdown() has become mark(), and markdownToHTML() has become mark_html().

Usage


renderMarkdown(file = NULL, output = NULL, ...)

markdownToHTML(
  file = NULL,
  output = NULL,
  ...,
  options = getOption("markdown.HTML.options"),
  title = NULL,
  stylesheet = getOption("markdown.HTML.stylesheet"),
  header = getOption("markdown.HTML.header"),
  template = getOption("markdown.HTML.template", TRUE),
  fragment.only = FALSE,
  encoding = "UTF-8"
)

Arguments

file, output, ..., options, template
Arguments to be passed to new functions.

title, stylesheet, header
Arguments to be passed to meta = list(title = , css = , header-includes = ), which is passed to mark_html().

fragment.only
Whether to generate a fragment or a full HTML document.

encoding
Ignored.

[Package markdown version 1.12 Index]