rendererOutputType {markdown}	R Documentation

Fetch the Renderer Output Type

Description

markdown allows up to seven renderers to be registered by users, and each must provide the type of output returned, either character or raw for binary output. HTML is provided by the package and outputs character.

Usage

rendererOutputType(name)

Arguments

name
a character string naming the renderer.

Value

The character string with a value of either character or raw.

See Also

markdownToHTML, registeredRenderers

Examples

# List all available renderers
rendererOutputType("HTML")

[Package markdown version 1.0 Index]