smartypants {markdown}	R Documentation

smartypants: ASCII punctuation to HTML entities

Description

smartypants transforms plain ASCII punctuation characters into smart typographic punctuation HTML entities.

Usage

smartypants(file, output, text)

Arguments

file
a character string giving the pathname of the file to read from. If it is omitted from the argument list, then it is presumed that the text argument will be used instead.

output
a character string giving the pathname of the file to write to. If it is omitted, then it is presumed that the user expects the results returned as a character string.

text
a character vector containing the markdown text to transform.

Value

smartypants returns NULL invisibly when output is to a file, and a character string otherwise.

See Also

markdownExtensions, markdownHTMLOptions, markdownToHTML.

For a description of the original markdown version:

http://daringfireball.net/projects/markdown/

The original Sundown library on github:

https://github.com/vmg/sundown

C stubs for writing new renders are in inst/include/markdown_rstubs.[ch].

Examples

cat(smartypants(text = "1/2 (c)\n"))

[Package markdown version 1.0 Index]